Chore: move unit test.
